Move update logic to update.js.
