+++ index.js
@@ -4,47 +4,31 @@ window.Phaser = require('phaser/build/custom/phaser-split')
 
 var Box = require('./box')
 var grid = require('./utils/grid')
+var updateModule = require('./update')
 
 var gameHeight = 950
 var gameWidth = 1200
 var game = new Phaser.Game(gameWidth, gameHeight, Phaser.CANVAS, 'phaser-example', { preload: preload, create: create, update: update });
 
 function preload() {
-
     game.load.image('atari', 'assets/block.png');
     game.load.image('background', 'assets/background2.png');
     game.load.spritesheet('dude', 'assets/dude.png', 32, 48);
-
 }
 
 var sprite;
-
-var player;
-var facing = 'left';
-var jumpTimer = 0;
-var cursors;
-var jumpButton;
-var bg;
-var yAxis = p2.vec2.fromValues(0, 1);
-
-
-
+var updateParameters = {}
 function create() {
-    bg = game.add.tileSprite(0, 0, gameWidth, gameHeight, 'background');
+    var bg = game.add.tileSprite(0, 0, gameWidth, gameHeight, 'background');
 
     //  Enable p2 physics
     game.physics.startSystem(Phaser.Physics.P2JS);
-
-    //player.body.collideWorldBounds = true;
-    //player.body.maxVelocity.y = 500;
-    //player.body.setSize(20, 32, 5, 16);
-
     game.physics.p2.gravity.y = 2500;
     game.physics.p2.world.defaultContactMaterial.friction = 0.3;
     game.physics.p2.world.setGlobalStiffness(1e5);
 
     //  Add a sprite
-    player = game.add.sprite(200, 200, 'dude');
+    var player = game.add.sprite(200, 200, 'dude');
     player.animations.add('left', [0, 1, 2, 3], 10, true);
     player.animations.add('turn', [4], 20, true);
     player.animations.add('right', [5, 6, 7, 8], 10, true);
@@ -62,12 +46,6 @@ function create() {
     //  4 trues = the 4 faces of the world in left, right, top, bottom order
     game.physics.p2.setWorldMaterial(worldMaterial, true, true, true, true);
 
-
-    //  A stack of boxes - you'll stick to these
-    // for (var i = 9; i >= 0; i--) {
-    //   box.placeBox(i, 9 - i)
-    // }
-
     //Creates sample grid
     for (var location of grid.boxLocations) {
       box.placeBox(location.y, location.x)
@@ -75,85 +53,24 @@ function create() {
 
     //  Here is the contact material. It's a combination of 2 materials, so whenever shapes with
     //  those 2 materials collide it uses the following settings.
-
     var groundPlayerCM = game.physics.p2.createContactMaterial(spriteMaterial, worldMaterial, { friction: 0.0 });
     var groundBoxesCM = game.physics.p2.createContactMaterial(worldMaterial, box.material, { friction: 0.6 });
 
-    //  Here are some more options you can set:
+    var cursors = game.input.keyboard.createCursorKeys();
+    var jumpButton = game.input.keyboard.addKey(Phaser.Keyboard.SPACEBAR);
+    var facing = "left";
 
-    // contactMaterial.friction = 0.0;     // Friction to use in the contact of these two materials.
-    // contactMaterial.restitution = 0.0;  // Restitution (i.e. how bouncy it is!) to use in the contact of these two materials.
-    // contactMaterial.stiffness = 1e3;    // Stiffness of the resulting ContactEquation that this ContactMaterial generate.
-    // contactMaterial.relaxation = 0;     // Relaxation of the resulting ContactEquation that this ContactMaterial generate.
-    // contactMaterial.frictionStiffness = 1e7;    // Stiffness of the resulting FrictionEquation that this ContactMaterial generate.
-    // contactMaterial.frictionRelaxation = 3;     // Relaxation of the resulting FrictionEquation that this ContactMaterial generate.
-    // contactMaterial.surfaceVelocity = 0.0;        // Will add surface velocity to this material. If bodyA rests on top if bodyB, and the surface velocity is positive, bodyA will slide to the right.
-
-
-    cursors = game.input.keyboard.createCursorKeys();
-    jumpButton = game.input.keyboard.addKey(Phaser.Keyboard.SPACEBAR);
-
-}
-
-function isJumping() {
-    return jumpButton.isDown || cursors.up.isDown;
+    console.log(player);
+    updateParameters = {
+        facing: facing,
+        player: player,
+        cursors: cursors,
+        jumpButton: jumpButton,
+        game: game
+    }
 }
 
 function update() {
-    var velocityAbs = 300;
-    player.body.velocity.x = 0;
-
-    if (cursors.left.isDown) {
-        player.body.velocity.x = -velocityAbs;
-
-        if (facing != 'left') {
-            player.animations.play('left');
-            facing = 'left';
-        }
-    } else if (cursors.right.isDown) {
-        player.body.velocity.x = velocityAbs;
-
-        if (facing != 'right')
-        {
-            player.animations.play('right');
-            facing = 'right';
-        }
-    } else { if (facing != 'idle') {
-            player.animations.stop();
-
-            if (facing == 'left') {
-                player.frame = 0;
-            } else {
-                player.frame = 5;
-            }
-
-            facing = 'idle';
-        }
-    }
-
-    if (isJumping() && checkIfCanJump()) {
-        player.body.velocity.y = -700; 
-    }
-
+    return updateModule(updateParameters)
 }
 
-function checkIfCanJump() {
-    var result = false;
-    for (var i=0; i < game.physics.p2.world.narrowphase.contactEquations.length; i++) {
-        var c = game.physics.p2.world.narrowphase.contactEquations[i];
-
-        if (c.bodyA === player.body.data || c.bodyB === player.body.data) {
-            var d = p2.vec2.dot(c.normalA, yAxis);
-
-            if (c.bodyA === player.body.data) {
-                d *= -1;
-            }
-
-            if (d > 0.5) {
-                result = true;
-            }
-        }
-    }
-
-    return result;
-}
